setView sets element.className to 'js-chat'
Note: I went and installed chai-jquery and jquery thinking that I
needed them to make this test pass. I was wrong. At this point in time
we are not even testing the real dom, we're just chilling in regular
javascript, which is why I commented out requiring those dependencies
in karma.conf