userFormController now takes username input of form
Also added was observable.coffee and a observable_test.coffee from
chapter 11. It's used with the handleSubmit callback method which
gets triggered when the 'submit' event from the browser is fired
inside of set view. handleSubmit itself then fires an event called
"user" that we observe in the tests and then assert that it's callback
gets called with the user's name passed to it